+++ livre/core/render/FrameInfo.cpp
@@ -23,9 +23,11 @@ namespace livre
 {
 
 FrameInfo::FrameInfo( const Frustum& frustum_,
-                      const uint32_t frameId_ )
+                      const uint32_t frameId_,
+                      const uint32_t frameCounter_ )
     : frustum( frustum_ )
     , frameId( frameId_ )
+    , frameCounter( frameCounter_ )
     , nAvailable( 0 )
     , nNotAvailable( 0 )
 {