+++ livre/eq/Window.cpp
@@ -69,10 +69,12 @@ public:
                         pipe->getFrameData()->getVRParameters().getMaxGPUCacheMemoryMB();
         _textureCache.reset( new TextureCache( node->getTextureDataCache(),
                                                maxGpuMemory * LB_1MB, GL_LUMINANCE8 ));
-
-        const size_t computeThreads = 4;
+        const size_t renderThreads = 2;
+        const size_t computeThreads = 2;
         const size_t uploadThreads = 4;
         _renderPipeline.reset( new RenderPipeline( *_textureCache,
+                                                   node->getHistogramCache(),
+                                                   renderThreads,
                                                    computeThreads,
                                                    uploadThreads,
                                                    _glContext ));