+++ livre/lib/pipeline/RenderPipeline.cpp
@@ -22,6 +22,7 @@
 #include <livre/lib/pipeline/VisibleSetGeneratorFilter.h>
 #include <livre/lib/pipeline/DataUploadFilter.h>
 #include <livre/lib/pipeline/RenderFilter.h>
+#include <livre/lib/pipeline/HistogramFilter.h>
 
 #include <livre/lib/cache/TextureCache.h>
 #include <livre/lib/cache/TextureDataCache.h>
@@ -35,11 +36,15 @@ namespace livre
 struct RenderPipeline::Impl
 {
     Impl( TextureCache& textureCache,
-          const size_t nComputeThreads,
+          HistogramCache& histogramCache,
+          const size_t nRenderThreads,
           const size_t nUploadThreads,
+          const size_t nComputeThreads,
           ConstGLContextPtr glContext )
         : _textureCache( textureCache )
+        , _histogramCache( histogramCache )
         , _dataSource( textureCache.getDataCache().getDataSource( ))
+        , _renderExecutor( nRenderThreads, glContext )
         , _computeExecutor( nComputeThreads, glContext )
         , _uploadExecutor( nUploadThreads, glContext )
         , _nUploadThreads( nUploadThreads )
@@ -59,6 +64,7 @@ struct RenderPipeline::Impl
                                                             i,
                                                             _nUploadThreads,
                                                             _textureCache );
+
             visibleSetGenerator.connect( "VisibleNodes",
                                          uploader, "VisibleNodes" );
             visibleSetGenerator.connect( "Params",
@@ -68,29 +74,41 @@ struct RenderPipeline::Impl
     }
 
     void createSyncPipeline( PipeFilter& renderFilter,
-                             Pipeline& computePipeline,
+                             PipeFilter& histogramFilter,
+                             Pipeline& renderPipeline,
                              Pipeline& uploadPipeline ) const
     {
         PipeFilter visibleSetGenerator =
-                computePipeline.add< VisibleSetGeneratorFilter >(
+                renderPipeline.add< VisibleSetGeneratorFilter >(
                     "VisibleSetGenerator", _dataSource );
 
         createAndConnectUploaders( uploadPipeline,
                                    visibleSetGenerator,
                                    renderFilter );
+
+        for( size_t i = 0; i < _nUploadThreads; ++i )
+        {
+            std::stringstream name;
+            name << "DataUploader" << i;
+            PipeFilter uploader =
+                    static_cast< const livre::PipeFilter& >(
+                        uploadPipeline.getExecutable( name.str( )));
+            uploader.connect( "CacheObjects", histogramFilter, "CacheObjects" );
+        }
     }
 
     void createAsyncPipeline( PipeFilter& renderFilter,
                               PipeFilter& redrawFilter,
-                              Pipeline& computePipeline,
+                              PipeFilter& histogramFilter,
+                              Pipeline& renderPipeline,
                               Pipeline& uploadPipeline ) const
     {
         PipeFilter visibleSetGenerator =
-                computePipeline.add< VisibleSetGeneratorFilter >(
+                renderPipeline.add< VisibleSetGeneratorFilter >(
                     "VisibleSetGenerator", _dataSource );
 
         PipeFilter renderingSetGenerator =
-                computePipeline.add< RenderingSetGeneratorFilter >(
+                renderPipeline.add< RenderingSetGeneratorFilter >(
                     "RenderingSetGenerator", _textureCache );
 
         visibleSetGenerator.connect( "VisibleNodes",
@@ -98,6 +116,9 @@ struct RenderPipeline::Impl
 
         renderingSetGenerator.connect( "CacheObjects",
                                       renderFilter, "CacheObjects" );
+
+        renderingSetGenerator.connect( "CacheObjects",
+                                      histogramFilter, "CacheObjects" );
 
         renderingSetGenerator.connect( "RenderingDone",
                                        redrawFilter, "RenderingDone" );
@@ -111,30 +132,39 @@ struct RenderPipeline::Impl
                  const FrameInfo& frameInfo,
                  const Range& dataRange,
                  const PixelViewport& pixelViewPort,
+                 const Viewport& viewport,
                  PipeFilter redrawFilter,
+                 PipeFilter sendHistogramFilter,
                  Renderer& renderer,
                  size_t& nAvailable,
                  size_t& nNotAvailable ) const
     {
-
         PipeFilterT< RenderFilter > renderFilter( "RenderFilter", _dataSource, renderer );
+        PipeFilterT< HistogramFilter > histogramFilter( "HistogramFilter", _histogramCache );
+        histogramFilter.getPromise( "Frustum" ).set( frameInfo.frustum );
+        histogramFilter.connect( "Histogram", sendHistogramFilter, "Histogram" );
+        histogramFilter.getPromise( "RelativeViewport" ).set( viewport );
+        sendHistogramFilter.getPromise( "RelativeViewport" ).set( viewport );
+        sendHistogramFilter.getPromise( "Id" ).set( frameInfo.frameCounter );
 
-        Pipeline computePipeline;
+        Pipeline renderPipeline;
         Pipeline uploadPipeline;
 
         if( vrParams.getSynchronousMode( ))
             createSyncPipeline( renderFilter,
-                                computePipeline,
+                                histogramFilter,
+                                renderPipeline,
                                 uploadPipeline );
         else
             createAsyncPipeline( renderFilter,
                                  redrawFilter,
-                                 computePipeline,
+                                 histogramFilter,
+                                 renderPipeline,
                                  uploadPipeline );
 
         PipeFilter visibleSetGenerator =
                 static_cast< const livre::PipeFilter& >(
-                    computePipeline.getExecutable( "VisibleSetGenerator" ));
+                    renderPipeline.getExecutable( "VisibleSetGenerator" ));
 
         visibleSetGenerator.getPromise( "Frustum" ).set( frameInfo.frustum );
         visibleSetGenerator.getPromise( "Frame" ).set( frameInfo.frameId );
@@ -146,10 +176,11 @@ struct RenderPipeline::Impl
         renderFilter.getPromise( "Viewport" ).set( pixelViewPort );
 
         if( !vrParams.getSynchronousMode( ))
-            redrawFilter.schedule( _computeExecutor );
-        computePipeline.schedule( _computeExecutor );
+            redrawFilter.schedule( _renderExecutor );
+        renderPipeline.schedule( _renderExecutor );
         uploadPipeline.schedule( _uploadExecutor );
-
+        sendHistogramFilter.schedule( _computeExecutor );
+        histogramFilter.schedule( _computeExecutor );
         renderFilter.execute();
 
         if( vrParams.getSynchronousMode( ))
@@ -162,7 +193,7 @@ struct RenderPipeline::Impl
         {
             const PipeFilter renderingSetGenerator =
                     static_cast< const livre::PipeFilter& >(
-                        computePipeline.getExecutable( "RenderingSetGenerator" ));
+                        renderPipeline.getExecutable( "RenderingSetGenerator" ));
 
             const UniqueFutureMap futures( renderingSetGenerator.getPostconditions( ));
             nAvailable = futures.get< size_t >( "AvailableCount" );
@@ -171,17 +202,23 @@ struct RenderPipeline::Impl
     }
 
     TextureCache& _textureCache;
+    HistogramCache& _histogramCache;
     const DataSource& _dataSource;
+    mutable SimpleExecutor _renderExecutor;
     mutable SimpleExecutor _computeExecutor;
     mutable SimpleExecutor _uploadExecutor;
     const size_t _nUploadThreads;
 };
 
 RenderPipeline::RenderPipeline( TextureCache& textureCache,
-                                const size_t nComputeThreads,
+                                HistogramCache& histogramCache,
+                                const size_t nRenderThreads,
                                 const size_t nUploadThreads,
+                                const size_t nComputeThreads,
                                 ConstGLContextPtr glContext )
     : _impl( new RenderPipeline::Impl( textureCache,
+                                       histogramCache,
+                                       nRenderThreads,
                                        nComputeThreads,
                                        nUploadThreads,
                                        glContext ))
@@ -194,7 +231,9 @@ void RenderPipeline::render( const VolumeRendererParameters& vrParams,
                              const FrameInfo& frameInfo,
                              const Range& dataRange,
                              const PixelViewport& pixelViewPort,
+                             const Viewport& viewport,
                              const PipeFilter& redrawFilter,
+                             const PipeFilter& sendHistogramFilter,
                              Renderer& renderer,
                              size_t& nAvailable,
                              size_t& nNotAvailable ) const
@@ -203,7 +242,9 @@ void RenderPipeline::render( const VolumeRendererParameters& vrParams,
                    frameInfo,
                    dataRange,
                    pixelViewPort,
+                   viewport,
                    redrawFilter,
+                   sendHistogramFilter,
                    renderer,
                    nAvailable,
                    nNotAvailable );