+++ doc/Changelog.md
@@ -3,6 +3,10 @@ Changelog {#Changelog}
 
 # master {#master}
 
+
+* [#301](https://github.com/BlueBrain/Livre/pull/301):
+  Histogram computation support is added to livre. The
+  histogram is also broadcasted with ZeroBuf.
 * [#303](https://github.com/BlueBrain/Livre/pull/303):
   Old renderer is revived for OSX. Fix #296
 * [#256](https://github.com/BlueBrain/Livre/pull/256):